+++ misc/mirv-script/package.json
@@ -10,14 +10,13 @@
 	},
 	"type": "module",
 	"scripts": {
-		"eslint": "eslint . --ext .ts --ignore-pattern=.gitignore,.eslintignore",
-		"build": "rollup -c",
+		"eslint": "eslint . --ext .ts,.mts --ignore-pattern=.gitignore,.eslintignore",
+		"build": "tsc --project tsconfig.mirv.json",
 		"build-scripts": "tsc --project tsconfig.scripts.json",
-		"server": "npm run build-scripts && node dist/server.js",
-		"client": "npm run build-scripts && node dist/client.js"
+		"server": "npm run build-scripts && node dist/node/server.js",
+		"client": "npm run build-scripts && node dist/node/client.js"
 	},
 	"devDependencies": {
-		"@rollup/plugin-typescript": "^11.1.6",
 		"@types/node": "^20.12.10",
 		"@types/ws": "^8.5.10",
 		"@typescript-eslint/eslint-plugin": "^7.8.0",
@@ -26,7 +25,6 @@
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-check-file": "^2.8.0",
 		"eslint-plugin-prettier": "^5.1.3",
-		"rollup": "^4.17.2",
 		"typescript": "^5.4.5"
 	},
 	"dependencies": {
