--- misc/mirv-script/rollup.config.mjs
@@ -1,9 +1,0 @@
-import typescript from '@rollup/plugin-typescript';
-
-export default {
-	input: 'src/index.ts',
-	output: {
-		file: 'dist/mirv.js'
-	},
-	plugins: [typescript()]
-};
