+++ misc/mirv-script/tsconfig.mirv.json
@@ -1,0 +1,8 @@
+{
+	"extends": "./tsconfig.json",
+	"compilerOptions": {
+		"outDir": "dist",
+		"noEmit": false
+	},
+	"include": ["src/index.mts", "src/types/*"]
+}
